requirements.txt
Flask==1.1.1
git+https://github.com/giacaglia/pytube.git
pydub==0.24.1
google-cloud-speech==1.3.2
google-api-python-client
Jinja2==2.10.1
gunicorn==19.9.0
certifi==2019.11.28
itsdangerous==1.1.0